requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
redis==5.0.0
python-dotenv==1.0.0
python-nanoid==2.0.0
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
